fixed get_prices, added inc_or_dec_prices
changed the way that getting prices/volumes work. now they just return a full trading day's worth of data back from the time given. I also added "inc_or_dec_prices" which gets the prices, does a linear fit, and returns 1 if the slope is positive and 0 if the slope is negative.